renamed package from adbms to sampleapp
