+++ sampleapp/src/main/scala/de/up/hpi/informationsystems/sampleapp/HelloWorld.scala
@@ -1,4 +1,4 @@
-package de.up.hpi.informationsystems.adbms
+package de.up.hpi.informationsystems.sampleapp
 
 object HelloWorld extends App {
   println("Hello World")
